# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.Error:
    properties:
      error:
        type: string
    type: object
  models.CustomIngredient:
    properties:
      custom_ingredient_id:
        minimum: 1
        type: integer
      quantity:
        type: number
    required:
    - custom_ingredient_id
    - quantity
    type: object
  models.Ingredient:
    properties:
      ingredient_id:
        minimum: 1
        type: integer
      quantity:
        type: number
    required:
    - ingredient_id
    - quantity
    type: object
  models.Recipe:
    properties:
      created_at:
        type: string
      custom_ingredients:
        items:
          $ref: '#/definitions/models.CustomIngredient'
        type: array
      ingredients:
        items:
          $ref: '#/definitions/models.Ingredient'
        type: array
      name:
        maxLength: 100
        minLength: 1
        type: string
      unit:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    - unit
    type: object
info:
  contact: {}
  description: SnackLog API for managing recipes created by users
  title: Recipe API for SnackLog
  version: "1.0"
paths:
  /recipe:
    get:
      description: Search recipes for the authenticated user by query string `q`.
        The query must be at least 3 characters long.
      parameters:
      - description: Search query (minimum 3 characters)
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Search recipes
      tags:
      - recipes
    post:
      consumes:
      - application/json
      description: Creates a new recipe for the authenticated user.
      parameters:
      - description: Recipe object
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a recipe
      tags:
      - recipe
  /recipe/{id}:
    delete:
      description: Deletes a recipe by ID for the authenticated user.
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete a recipe
      tags:
      - recipe
    get:
      description: Retrieves a recipe by ID for the authenticated user.
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      security:
      - ApiKeyAuth: []
      summary: Get a recipe by ID
      tags:
      - recipe
    put:
      consumes:
      - application/json
      description: Updates an existing recipe by ID for the authenticated user.
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: integer
      - description: Recipe object
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      security:
      - ApiKeyAuth: []
      summary: Update a recipe
      tags:
      - recipe
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"